+++ contrib/storage-np/src/main/java/org/apache/drill/exec/store/np/read/NPReaderFactory.java
@@ -1,13 +1,11 @@
 package org.apache.drill.exec.store.np.read;
 
-import org.apache.drill.common.expression.SchemaPath;
 import org.apache.drill.exec.physical.impl.scan.framework.ManagedReader;
 import org.apache.drill.exec.physical.impl.scan.framework.ManagedScanFramework;
 import org.apache.drill.exec.physical.impl.scan.framework.SchemaNegotiator;
 import org.apache.drill.exec.store.np.scan.NPSubScan;
 
 import java.util.Iterator;
-import java.util.List;
 import java.util.stream.Stream;
 
 /**
@@ -39,7 +37,7 @@ public class NPReaderFactory implements ManagedScanFramework.ReaderFactory {
                     ReaderProperties props = new ReaderProperties(
                             subScan.getScanSpec().getPluginConfig().getConnection(),
                             subScan.getColumns(),
-                            subScan.getFilters(), // add tablet filters here
+                            addTabletQueryRangeTo(subScan.getFilters(), tablet.getFilter()),
                             subScan.getScanSpec().getTableName(),
                             subScan.getSubScanId()
                     );
@@ -47,6 +45,30 @@ public class NPReaderFactory implements ManagedScanFramework.ReaderFactory {
                     return new NPBatchReader(props);
                 })
                 .iterator();
+    }
+
+    /**
+     * Decides how to build the final OJAI query to be executed by the corresponding reader.
+     *
+     * @param filter           OJAI query built from the query filters (WHERE clause).
+     * @param tabletQueryRange OJAI query that identify the tablet being query (OJAI Testing is {} since it is in
+     *                         memory and there are not tables.
+     * @return Final OJAI query (AND) to added as OJAI where to the query to be executed.
+     */
+    private String addTabletQueryRangeTo(String filter, String tabletQueryRange) {
+        if (!filter.equals("{}") && !tabletQueryRange.equals("{}")) {
+            return String.format("{\"$and\":[%s,%s]}", filter, tabletQueryRange);
+        }
+
+        if (filter.equals("{}")) {
+            return tabletQueryRange;
+        }
+
+        if (tabletQueryRange.equals("{}")) {
+            return filter;
+        }
+
+        return "";
     }
 
     @Override
@@ -67,42 +89,3 @@ public class NPReaderFactory implements ManagedScanFramework.ReaderFactory {
     }
 }
 
-class ReaderProperties {
-    private final String connectionString;
-    private final List<SchemaPath> projections;
-    private final String ojaiJsonFilters;
-    private final String tableName;
-    private final Integer subScanId;
-
-    public ReaderProperties(String connectionString,
-                            List<SchemaPath> projections,
-                            String ojaiJsonFilters,
-                            String tableName, Integer subScanId) {
-
-        this.connectionString = connectionString;
-        this.projections = projections;
-        this.ojaiJsonFilters = ojaiJsonFilters;
-        this.tableName = tableName;
-        this.subScanId = subScanId;
-    }
-
-    public String getConnectionString() {
-        return connectionString;
-    }
-
-    public List<SchemaPath> getProjections() {
-        return projections;
-    }
-
-    public String getOjaiJsonFilters() {
-        return ojaiJsonFilters;
-    }
-
-    public String getTableName() {
-        return tableName;
-    }
-
-    public Integer getSubScanId() {
-        return subScanId;
-    }
-}
